+++ packages/gui/electron-builder.config.cjs
@@ -1,5 +1,41 @@
+const { spawnSync } = require('node:child_process')
+
 const publishUrl = process.env.VUE_APP_PUBLISH_URL
 const publishProvider = process.env.VUE_APP_PUBLISH_PROVIDER
+
+function hasExecutable (command, args = ['--version']) {
+  const result = spawnSync(command, args, { stdio: 'ignore' })
+  return !result.error && result.status === 0
+}
+
+const linuxTargets = [
+  {
+    target: 'deb',
+    arch: ['x64', 'arm64', 'armv7l'],
+  },
+  {
+    target: 'AppImage',
+    arch: ['x64', 'arm64', 'armv7l'],
+  },
+  {
+    target: 'tar.gz',
+    arch: ['x64', 'arm64', 'armv7l'],
+  },
+]
+
+if (hasExecutable('rpmbuild')) {
+  linuxTargets.push({
+    target: 'rpm',
+    arch: ['x64', 'arm64', 'armv7l'],
+  })
+}
+
+if (hasExecutable('flatpak') && hasExecutable('flatpak-builder')) {
+  linuxTargets.push({
+    target: 'flatpak',
+    arch: ['x64'],
+  })
+}
 
 /** @type {import('electron-builder').Configuration} */
 module.exports = {
@@ -41,6 +77,24 @@ module.exports = {
     {
       from: 'extra',
       to: 'extra',
+      filter: [
+        '**/*',
+        '!xray/**',
+      ],
+    },
+    {
+      from: 'extra/xray',
+      to: 'extra/xray',
+      filter: [
+        '*.mmdb',
+        'geoip.dat',
+        'geosite.dat',
+      ],
+    },
+    {
+      // eslint-disable-next-line no-template-curly-in-string
+      from: 'extra/xray/${os}/${arch}',
+      to: 'extra/xray',
     },
   ],
   afterPack: './pkg/after-pack.cjs',
@@ -62,28 +116,7 @@ module.exports = {
   },
   linux: {
     icon: 'build/mac/',
-    target: [
-      {
-        target: 'deb',
-        arch: ['x64', 'arm64', 'armv7l'],
-      },
-      {
-        target: 'AppImage',
-        arch: ['x64', 'arm64', 'armv7l'],
-      },
-      {
-        target: 'tar.gz',
-        arch: ['x64', 'arm64', 'armv7l'],
-      },
-      {
-        target: 'rpm',
-        arch: ['x64', 'arm64', 'armv7l'],
-      },
-      {
-        target: 'flatpak',
-        arch: ['x64'],
-      },
-    ],
+    target: linuxTargets,
     appId: 'cn.docmirror.DevSidecar',
     category: 'System',
   },